src/server/pyproject.toml
[tool.poetry]
name = "legarda"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask-restplus = "^0.13.0"
flask-migrate = "^2.5"
flask-script = "^2.0"
flask-sqlalchemy = "^2.4"
flask-jwt-extended = "^3.24"
flask-bcrypt = "^0.7.1"
flask-testing = "^0.7.1"
psycopg2 = "^2.8"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
